services/notification-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pika==1.3.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
